+++ example/radar/radar_radolan_cdc.py
@@ -45,6 +45,7 @@ real-time for Germany.
 - Daily: https://docs.wradlib.org/en/stable/notebooks/radolan/radolan_showcase.html#RADOLAN-SF-Product # noqa
 """
 import logging
+import os
 
 import matplotlib.pyplot as pl
 import numpy as np
@@ -140,7 +141,7 @@ def radolan_grid_example():
         end_date="2020-09-04T12:00:00",
     )
 
-    for item in radolan.collect_data():
+    for item in radolan.query():
 
         # Decode data using wradlib.
         log.info("Parsing RADOLAN_CDC composite data for %s", item.timestamp)
@@ -151,7 +152,8 @@ def radolan_grid_example():
 
         # Plot and display data.
         plot(data, attributes, label)
-        pl.show()
+        if "PYTEST_CURRENT_TEST" not in os.environ:
+            pl.show()
 
 
 def main():